full update

+++ supa.js
@@ -17,19 +17,41 @@ async function getAllDokans({ dokanArea = null, dno = null }) {
     // console.log("index");
     dokans?.map((dokan) => {
       dokanDetails.push({
+        id: dokan.id,
         name: dokan.name,
         mobile: dokan.mobile,
         email: dokan.email,
       });
     });
 
-    dokanDetails.map((dokan) => {
+    dokanDetails.forEach(async (dokan) => {
       const dokanItem = document.createElement("div");
+      const dueTotal = await dokanDueTotal(dokan.name);
+
       // div.className="span-menu";
       dokanItem.style.display = "flex";
       dokanItem.style.flexWrap = "nowrap";
-      dokanItem.style.padding = "0.2rem";
-      dokanItem.innerHTML = `<div class="span-menu1">${dokan.name}</div><div class="span-menu1">${dokan.mobile}</div><div class="span-menu1">${dokan.email}</div>`;
+      dokanItem.style.padding = "0";
+      dokanItem.style.gap = "0.1rem";
+
+      dokanItem.innerHTML = `<div class="span-menu1">${dokan.name}</div>
+      <div class="span-menu1">${dokan.mobile}</div>
+      <div class="span-menu1">${dokan.email}</div>
+      <div id="dokan-delete-${
+        dokan.id
+      }" class="dokan-delete span-menu1" style="display:flex;align-items:center">
+      <span style="background-color:red;border-radius:1.8rem;text-align:center;border:2px solid black;cursor:pointer;padding:0.8rem">delete</span>
+      </div>
+      <div id="dokan-show-${
+        dokan.id
+      }" class="dokan-show span-menu1" style="display:flex;align-items:center">
+
+      <span style="background-color:green;border-radius:1.8rem;text-align:center;border:2px solid black;cursor:pointer;padding:0.8rem">view<span/></div>
+      <div class="span-menu1" id="dokan-due-${
+        dokan.id
+      }">${dueTotal} gm<p style='background-color:yellow;border-radius:1.5rem;font-size:small;color:black;text-align:center;margin-top:0.5rem'>${
+        dueTotal < 0 ? "to pay" : "to be received"
+      }</p></div>`;
 
       dokanArea?.append(dokanItem);
     });
@@ -46,6 +68,20 @@ async function getAllDokans({ dokanArea = null, dno = null }) {
       dno?.appendChild(option);
     });
   }
+}
+
+//Total Due
+async function dokanDueTotal(name) {
+  let total = 0;
+  const { data, error } = await supabase
+    .from("gold")
+    .select("due")
+    .eq("shop_name", name);
+  data?.forEach((el) => {
+    // console.log(due);
+    total = total + el.due;
+  });
+  return total.toFixed(2);
 }
 
 //add-dokans
@@ -155,14 +191,31 @@ async function getDokan() {
 
 //delete record
 async function deleteRecord(e) {
-  console.log("del");
-  const btn = e.target;
-  const id = btn.id;
-  console.log(id);
+  if (
+    window.location.pathname == "/dokan/transaction.html" ||
+    window.location.href.includes("transaction")
+  ) {
+    console.log("del");
+    const btn = e.target;
+    const id = btn.id;
+    console.log(id);
 
-  const { error } = await supabase.from("gold").delete().eq("id", id);
-  console.log(error);
-  document.querySelector(`button[id="${id}"]`).parentElement.remove();
+    const { error } = await supabase.from("gold").delete().eq("id", id);
+    console.log(error);
+    document.querySelector(`button[id="${id}"]`).parentElement.remove();
+  } else if (
+    window.location.pathname == "/dokan/dokans.html" ||
+    window.location.href.includes("dokans")
+  ) {
+    console.log(e);
+    const div = e.target;
+    const id = div.id;
+    const { error } = await supabase
+      .from("dokan")
+      .delete()
+      .eq("id", id.split("-")[2]);
+    document.querySelector(`div[id="${id}"]`).parentElement.remove();
+  }
 }
 
 //filter
@@ -313,6 +366,11 @@ if (window.location.href.includes("transaction")) {
 if (window.location.href.includes("dokans")) {
   const dokanArea = document.getElementById("dokan-area");
   getAllDokans({ dokanArea });
+  document.addEventListener("click", function (e) {
+    if (e.target.classList.contains("dokan-delete")) {
+      deleteRecord(e);
+    }
+  });
 }
 
 //---------------------------------------------------
